requirements.txt
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
openai>=1.0.0
python-dotenv>=1.0.0

